requirements.txt
Django==1.10b1

boto==2.40
ipython==5.0.0b3
piprot==0.9.6
psycopg2==2.6.1
raven==5.20.0

django-axes==2.0.0
django-compressor==2.0
django-extensions==1.6.7
django-jenkins==0.18.1
django-markdownx==1.6
django-storages==1.4.1
django-countries==3.4.1
django-registration==2.1.2
